+++ spotify/manager/session.py
@@ -59,14 +59,6 @@ class SpotifySessionManager(object):
 
     disconnect = terminate
 
-    def wake(self, session=None):
-        """
-        This is called by the Spotify subsystem to wake up the main loop.
-        """
-        if self.timer is not None:
-            self.timer.cancel()
-        self.awoken.set()
-
     def logged_in(self, session, error):
         """
         Callback.
@@ -136,9 +128,13 @@ class SpotifySessionManager(object):
         """
         Callback.
 
-        When this method is called, you should make sure that
-        :meth:`session.process_events() <spotify.Session.process_events>` is
-        called.
+        When this method is called by ``libspotify``, one should call
+        :meth:`session.process_events() <spotify.Session.process_events>`.
+
+        If you use the :class:`SessionManager`'s default loop, the default
+        implementation of this method does the job. Though, if you implement
+        your own loop for handling Spotify events, you'll need to override this
+        method.
 
         .. warning::
             This method is called from an internal thread in libspotify. You
@@ -148,7 +144,9 @@ class SpotifySessionManager(object):
         :param session: the current session.
         :type session: :class:`spotify.Session`
         """
-        pass
+        if self.timer is not None:
+            self.timer.cancel()
+        self.awoken.set()
 
     def music_delivery(self, session, frames, frame_size, num_frames,
             sample_type, sample_rate, channels):